+++ dev/backend/app.py
@@ -2,6 +2,8 @@ from flask import Flask, jsonify, request, send_file
 from flask_cors import CORS
 from flask_socketio import SocketIO
 
+from train.train_cartpole import train_cartpole
+
 from routes import setup_routes_base, setup_db
 from sockets import setup_sockets
 
@@ -14,6 +16,11 @@ setup_routes_base(app=app)
 setup_db(app=app)
 
 setup_sockets(socketio=socketio)
+
+@socketio.on('CartPole')
+def train_CartPole(datas):
+    global thread
+    train_cartpole(datas, socketio)
 
 if __name__ == '__main__':
     socketio.run(app, debug=True, host='0.0.0.0', port=5000)
